Module spirit::fragment::driver

A collection of Drivers for use by Fragments.

This module contains the Driver trait, several implementations of that trait for general use and some support types around them.

Structs

CacheEq
A Driver that caches the Resource if the Fragment doesn't change at all.

CacheId
ID used by the Driver to identify specific instance of the resource.

CacheSimilar
A Driver that caches both created Resources and their Seeds to create as little as possible.

IdGen
Generator of IDs for the Driver.

IdMapping
An object to map Instructions from multiple drivers into instruction sequences not containing duplicate IDs.

OnceDriver
A Driver for a single-shot initialization.

RefDriver
An adaptor Driver for references.

SeqDriver
A plumbing Driver for sequences of fragments.

Trivial
A trivial Driver that does no caching at all.

Enums

Comparison
A result of the Comparable trait.

Instruction
One instruction from the Driver.

Traits

Comparable
Fragments that can be compared for similarity.

Driver
The Driver of a Pipeline.