Struct spirit::fragment::driver::RefDriver

pub struct RefDriver<Inner>(_);

An adaptor Driver for references.

This is used behind the scenes to wrap a driver for F to create a driver for &F.

Implementations

impl<Inner> RefDriver<Inner>

pub fn new(inner: Inner) -> Self

Creates the driver.

It is available to support also drivers that are provided and created by the user. Usually, the Default implementation is used within the Fragment.