Module spirit::fragment::pipeline

The home of the Pipeline.

The high level overview of what a Pipeline is and how it works is in the fragment module.

The rest of the things here is mostly support plumbing and would optimally be hidden out of the public interface, but it needs to be public due to limitations of Rust. The user doesn't need to care much about the other types. Despite all that, the types are not hidden from the documentation to provide some guidance and clickable links.

Structs

AndThen
A Transformation that maps a Resource through a fallible closure.

CfgExtractor
A wrapper to turn a FnMut(Config) -> R into an Extractor.

ChainedTransformation
A wrapper that composes two Transformations into one.

CompiledPipeline
An internal intermediate type.

Map
A Transformation than maps a Resource through a closure.

MultiError
An error caused by multiple other errors.

NopTransformation
A Transformation that does nothing.

Pipeline
The Pipeline itself.

SetInstaller
A Transformation that replaces the Installer of a pipeline.

Traits

BoundedCompiledPipeline
Trait alias for one concrete lifetime of a Pipeline.