Trait spirit::fragment::pipeline::BoundedCompiledPipeline

pub trait BoundedCompiledPipeline<'a, O, C> {
    fn run(
        me: &Arc<Mutex<Self>>, 
        opts: &'a O, 
        config: &'a C
    ) -> Result<Action, Vec<AnyError>>;
}

Trait alias for one concrete lifetime of a Pipeline.

Pipelines are fed with only references to the configuration and command line options and a lot of the processing can happen through references only. As a result, most of the trait bounds in around the pipelines are HRTBs.

This is an internal trait alias, describing the Pipeline bounds for a single concrete lifetime. This makes the bounds of the Extension implementation actually almost manageable instead of completely crackpot insane.

However, as the user is not able to get the hands on any instance implementing this trait, it is quite useless and is public only through the trait bounds.

Required methods

fn run(
me: &Arc<Mutex<Self>>,
opts: &'a O,
config: &'a C
) -> Result<Action, Vec<AnyError>>

Performs one iteration of the lifetime.

Loading content...