Module spirit::prelude

The prelude.

To use the spirit libraries effectively, a lot of traits need to be imported. Instead of importing them one by one manually, the prelude contains the most commonly used imports that are used around application runtime management. This imports only traits and only in anonymous mode (eg. pub use spirit::SpiritBuilder as _).

This can be imported as use spirit::prelude::*.

Re-exports

pub use super::ConfigBuilder as _;

pub use super::Extensible as _;

pub use super::Fragment as _;

Traits

_
An interface to turn the spirit Builder into a Spirit and possibly run it.