Struct spirit::Builder

#[must_use = "The builder is inactive without calling `run` or `build`"]pub struct Builder<O = Empty, C = Empty> { /* fields omitted */ }

The builder of Spirit.

This is returned by the Spirit::new.

Interface

Most of the methods are available through the following traits. You want to bring them into scope (possibly by use spirit::prelude::*) and look at their methods.

ConfigBuilder to specify how configuration should be loaded.

Extensible to register callbacks.

SpiritBuilder to turn the builder into Spirit.

Implementations

impl<O, C> Builder<O, C> where
C: DeserializeOwned + Send + Sync + 'static,
O: StructOpt + Sync + Send + 'static,

pub fn config_loader(self, loader: CfgBuilder) -> Self

Sets the inner config loader.

Instead of configuring the configuration loading on the spirit Builder, the inner config loader Builder can be set directly.