Function spirit::utils::serialize_duration

pub fn serialize_duration<S: Serializer>(
    dur: &Duration, 
    s: S
) -> Result<S::Ok, S::Error>

Serialize a duration.

This can be used in configuration structures containing durations. See deserialize_duration for the counterpart.

The default serialization produces human unreadable values, this is more suitable for dumping configuration users will read.

Examples

use std::time::Duration;

use serde::{Deserialize, Serialize};

#[derive(Clone, Debug, Eq, PartialEq, Deserialize, Serialize)]
struct Cfg {
    #[serde(
        serialize_with = "spirit::utils::serialize_duration",
        deserialize_with = "spirit::utils::deserialize_duration",
    )]
    how_long: Duration,
}